adding post game and moving checks into new namespace for cleanup
