+++ StateLibrary/Actions/EventChecks/ScoreCheck.cs
@@ -1,0 +1,13 @@
+﻿using DomainObjects;
+using StateLibrary.Interfaces;
+
+namespace StateLibrary.Actions.EventChecks
+{
+    public sealed class ScoreCheck : IGameAction
+    {
+        public void Execute(Game game)
+        {
+            
+        }
+    }
+}
