+++ StateLibrary/Actions/PostGame.cs
@@ -3,11 +3,11 @@ using StateLibrary.Interfaces;
 
 namespace StateLibrary.Actions
 {
-    public sealed class ScoreCheck : IGameAction
+    public class PostGame : IGameAction
     {
         public void Execute(Game game)
         {
-            
+
         }
     }
 }
