+++ StateLibrary/StateLibrary.csproj
@@ -12,8 +12,4 @@
     <ProjectReference Include="..\DomainObjects\DomainObjects.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Actions\EventChecks\" />
-  </ItemGroup>
-
 </Project>
